Re-add rbac.go to deal with app-wide rbac
Specifically add a service account to be used by all components
that don't require one for other reasons.
